ADD admin

+++ TP4/Inscripcion/InscripcionesPorAlumno.cs
@@ -12,7 +12,7 @@ namespace TP4
     {
         //public int NInscripcion { get; set; }
         public int NRegistro { get; set; }
-        public int RankingAlumno { get; set; }
+        public double RankingAlumno { get; set; }
         public int CodigoMateria { get; set; }
         public string NombreMateria { get; set; }
         public int CantidadInscriptos { get; set; }
@@ -23,7 +23,7 @@ namespace TP4
         {
             var datos = linea.Split('-');
             NRegistro = int.Parse(datos[0]);
-            RankingAlumno = int.Parse(datos[1]);
+            RankingAlumno = double.Parse(datos[1]);
             CodigoMateria = int.Parse(datos[2]);
             NombreMateria = (datos[3]);
         }
@@ -32,7 +32,7 @@ namespace TP4
 
         public static List<InscripcionesPorAlumno> inscripcionesPorAlumno = new List<InscripcionesPorAlumno>();
 
-        public static void EscribirInscripcionEnTXT(int numRegistro, int codMateria, string nomMateria, int rankingAlumno)
+        public static void EscribirInscripcionEnTXT(int numRegistro, int codMateria, string nomMateria, double rankingAlumno)
         {
             string fileName = "TP4/TXT/InscripcionesPorAlumnos.txt";
             string basePath = Environment.CurrentDirectory;
@@ -54,7 +54,7 @@ namespace TP4
         }
 
 
-        public static void AgregarInscripcion(int numRegistro, int codMateria, string nomMateria, int rankingAlumno)
+        public static void AgregarInscripcion(int numRegistro, int codMateria, string nomMateria, double rankingAlumno)
         {
             inscripcionesPorAlumno.Add(new InscripcionesPorAlumno()
             {
